code reorganize
