+++ bot/handlers/callback_handlers.py
@@ -1,5 +1,9 @@
-from aiogram import types
+from aiogram import types, Router, Dispatcher
+from aiogram.types import CallbackQuery
 
+from bot.common.bot_crud import get_user_info
+
+router = Router()
 callback_handlers = {}
 
 
@@ -9,6 +13,19 @@ def register_callback_handler(callback_prefix: str):
         return func
 
     return decorator
+
+
+@router.callback_query()
+async def handle_callbacks(callback: CallbackQuery):
+    telegram_id = callback.from_user.id
+    language_code = get_user_info(telegram_id).language_code
+
+    for prefix, handler in callback_handlers.items():
+        if callback.data.startswith(prefix):
+            await handler(callback, language_code)
+            return
+
+    await callback.answer("Unknown action")
 
 
 @register_callback_handler("edit_profile")
@@ -32,3 +49,7 @@ async def start_edit_profile(callback: types.CallbackQuery, language_code: str):
     )
     await callback.message.answer(instruction)
     await callback.answer()
+
+
+def register_callback_handlers(dispatcher: Dispatcher):
+    dispatcher.include_router(router)
